RMRKNesting partial rewrite
Makes explicit difference when minting vs minting to NFT, no longer deciding according to passed data.

From @Ayuilos :
First of all, Nesting should be an ERC721 extension that implements the nesting function, then its usage scenarios are roughly 2:
- was downgraded to pure ERC721 to use
- Use as Nesting token
Then when it is used as a Nesting token, in theory, it should refer to the ERC721 interface to achieve almost the same purpose, such as the safe transfer and safe mint of the token as a Nesting, I think the “safe” here refers to: if the target address is a contract, then it is necessary to check whether the target contract is ready to receive the token. The “ready” refers to whether the target contract can use all the functions of the token to be received. Therefore, the sefe method of Nesting needs to be distinguished from the sefe method of ERC721 , which is why I added _safeMintNesting and _safeTransferNestingFrom.