Extract errors to a library
Errors were extractred to a library smart contract. The contract option
was picked over the library simply because deployment in such manner
should conserve gas. The errors library smart contracts was then
imported in the smart contracts and the inheritance tree respected, so
that each instance only gets it inherited once.