Extract errors to a library and allows to add resources to not yet existing tokens (#94)
Errors were extractred to a library smart contract. The contract option
was picked over the library simply because deployment in such manner
should conserve gas. The errors library smart contracts was then
imported in the smart contracts and the inheritance tree respected, so
that each instance only gets it inherited once.

The errors contained within the RMRKErrors have been documented using
NatSpec. The shorthand (no definition) was used for notice type of
comments to keep the library as readable as possible.

Allows to add resources to not yet existing tokens. This affects mocks and implementations, not the core.