Soulbound extensions (#163)
Implements 3 variants of soulbound extension.

RMRKSoulboundAfterBlock: On deploy, the last block where tokens can be transferred is set. After that, every token will be Soulbound.
RMRKSoulboundAfterTransactions: On deploy, the max number of transactions per token is set. After a token has been transferred this number of times, it becomes Soulbound.
RMRKSoulboundPerToken: Tokens are set Soulbound or not individually.

---------

Co-authored-by: Jan Turk <jan.turk12@gmail.com>